refactoring component folder and add event onclick
