+++ src/components/CartWidget/index.js
@@ -1,0 +1,2 @@
+import CartWidget from "./CartWidget";
+export default CartWidget;
