+++ src/components/ItemListContainer/index.js
@@ -1,0 +1,2 @@
+import ItemListContainer from "./ItemListContainer";
+export default ItemListContainer;
