+++ src/components/NavBar/NavBar.jsx
@@ -3,13 +3,12 @@ import Nav from "react-bootstrap/Nav";
 import Navbar from "react-bootstrap/Navbar";
 import NavDropdown from "react-bootstrap/NavDropdown";
 import "./NavBar.css";
-import CartWidget from "./CartWidget";
+import CartWidget from "../CartWidget/CartWidget";
 
 function BasicExample() {
   return (
     <Navbar fixed="top" expand="lg" className={["bg-body-tertiary, navbar"]}>
       <Container>
-        {/* logo */}
         <Navbar className="bg-body-tertiary">
           <Container>
             <Navbar.Brand href="#home">
@@ -24,19 +23,30 @@ function BasicExample() {
             <Nav.Link href="#home">Home</Nav.Link>
             <Nav.Link href="#link">Link</Nav.Link>
             <NavDropdown title="Dropdown" id="basic-nav-dropdown">
-              <NavDropdown.Item href="#action/3.1">Action</NavDropdown.Item>
-              <NavDropdown.Item href="#action/3.2">Another action</NavDropdown.Item>
-              <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
+              <NavDropdown.Item onClick={onItemClicked} href="#action/3.1">
+                Action
+              </NavDropdown.Item>
+              <NavDropdown.Item onClick={onItemClicked} href="#action/3.2">
+                Another action
+              </NavDropdown.Item>
+              <NavDropdown.Item onClick={onItemClicked} href="#action/3.3">
+                Something
+              </NavDropdown.Item>
               <NavDropdown.Divider />
-              <NavDropdown.Item href="#action/3.4">Separated link</NavDropdown.Item>
+              <NavDropdown.Item onClick={onItemClicked} href="#action/3.4">
+                Separated link
+              </NavDropdown.Item>
             </NavDropdown>
           </Nav>
         </Navbar.Collapse>
-        {/* cartwidget */}
         <CartWidget />
       </Container>
     </Navbar>
   );
 }
+
+const onItemClicked = () => {
+  alert("Item clicked");
+};
 
 export default BasicExample;
